+++ src/renderer/src/view/Terminals/RightDrawer/FlowControlView.tsx
@@ -3,7 +3,6 @@ import { observer } from 'mobx-react-lite';
 import ArrowDownwardIcon from '@mui/icons-material/ArrowDownward';
 
 import { App } from '@/model/App';
-import { BASIC_TOOLTIP_SETTINGS } from '@/view/SharedConfig';
 import { PortState } from '@/model/Settings/PortSettings/PortSettings';
 import { CustomAccordionSummary } from './CustomAccordionSummary';
 
@@ -71,7 +70,7 @@ export default observer((props: Props) => {
 
           {/* Row 1: RTS */}
           <Tooltip
-            {...BASIC_TOOLTIP_SETTINGS}
+            {...app.settings.displaySettings.getBasicTooltipConfig()}
             title="Request To Send. Write only. Traditionally asserted by the DTE (e.g. computer) to the DCE (e.g. modem) to indicate that it is ready to send data. Commonly used for hardware flow control."
             placement="left"
           >
@@ -96,7 +95,7 @@ export default observer((props: Props) => {
 
           {/* Row 2: CTS */}
           <Tooltip
-            {...BASIC_TOOLTIP_SETTINGS}
+            {...app.settings.displaySettings.getBasicTooltipConfig()}
             title="Clear To Send. Read/write. Commonly used for hardware flow control. If you set this and it quickly changes, it means the other end is driving the signal."
             placement="left"
           >
@@ -121,7 +120,7 @@ export default observer((props: Props) => {
 
           {/* Row 3: DTR */}
           <Tooltip
-            {...BASIC_TOOLTIP_SETTINGS}
+            {...app.settings.displaySettings.getBasicTooltipConfig()}
             title="Data Terminal Ready. Write only. Traditionally asserted by the DTE when it was ready to receive data. Not as commonly used for flow control as RTS/CTS. Some older devices used DTR and RTS as a source of power."
             placement="left"
           >
@@ -146,7 +145,7 @@ export default observer((props: Props) => {
 
           {/* Row 4: DSR */}
           <Tooltip
-            {...BASIC_TOOLTIP_SETTINGS}
+            {...app.settings.displaySettings.getBasicTooltipConfig()}
             title="Data Set Ready. Read/write. Traditionally sent by the DCE (e.g. modem) to the DTE (e.g. computer) to indicate that the DCE is ready to receive data. If you set this and it quickly changes, it means the other end is driving the signal."
             placement="left"
           >
@@ -171,7 +170,7 @@ export default observer((props: Props) => {
 
           {/* Row 5: DCD (read-only) */}
           <Tooltip
-            {...BASIC_TOOLTIP_SETTINGS}
+            {...app.settings.displaySettings.getBasicTooltipConfig()}
             title="Data Carrier Detect. Read only. Was traditionally driven high (asserted, 1) by the modem (DCE) when it detected a carrier signal, which meant it was connected to another modem on the other end."
             placement="left"
           >